feat(test-utils): introduce test-utils
Some functions in interface-manager and mgmt are (currently) only used in tests.
While these functions should not be part of any public api, they are needed by the test suites of more than one package in this workspace.
This commit gathers these functions into a test-utils crate so that we can use them where needed without cluttering our public api.

This step is necessary for the VPC manager integration to prevent test regression.
Once the VPC manager is integrated, it will cause extant tests to start creating linux network interfaces.
These tests will then fail in the absence of the privilege escalation features available in the (new) test-utils package.

Thus, our choices are to either duplicate the test code, or factor it out into a common library.
This commit takes the latter approach.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>