feat(mgmt): make build internal config fallible
The previous code assumed that building the internal config could
not fail as such a step happens once the external config has been
validated. When building the internal config, however, some values
have to be chosen, which may fail to satisfy constraints
(e.g. uniqueness) required later on. To solve this, let the
building of the internal config be fallible.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>