feat(net): change vxlan encapsulation API
To encapsulate a packet in VxLAN, a UDP header needs to be built.
In the current API, the caller is requested to build and pass the
UDP information. All of the information in the UDP header, however,
can be automatically be set by the API. This includes:
  * dst port, which should almost always be the well-known value.
  * source port, which can be computed from the payload itself.
  * length
  * checksums (optional for ipv4)

This patch changes the API to not require UDP header and build it
automatically, so that callers need not have to set those.
This may be useful when needing to encapsulate traffic e.g. for
L2 EVIs.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>