feat(ci): use nextest to find flaky tests
Requires tests (especially fuzz tests) to fail 3 times before marking them as truly failed.

The theory is that this

1. Helps debug failing tests by giving the developer more examples of failures,
2. Does not block CI when the fuzzer finds new (but rare) bugs in functionality unrelated to the current PR.

This way we don't need to static seed (which prevents the fuzzer from truly exploring new territory), nor do we need to skip fuzz tests in CI (which would be highly unfortunate).

Signed-off-by: Daniel Noland <daniel@githedgehog.com>