fix(ci): sleep hack
We seem to be triggering a rate limit by launching (at nearly the same time) three different jobs which all invoke binstall.

This is crude, but we can hack around the rate limit by adding a trivial delay depending on the rust version.

I hate this, but I hate it less than random 3 minute delays in CI.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>