test(routing): Re-implement From<&str> trait for Prefix
The trait was first implemented in commit 39abac4465e9 ("feat(routing):
Implement From<&str> trait for Prefix"). It was later removed in commit
4caec75ebb02 ("feat(routing)!: Add TryFrom for PrefixString and
expect_from for tests"), seemingly because the introduction of the
PrefixString struct made it less relevant to build prefixes from
strings. But it remains faster, and easier to read, to use the
"1.2.3.4/24".into() notation to build numerous prefixes for unit tests,
so let's re-add support for it.

Signed-off-by: Quentin Monnet <qmo@qmon.net>