feat(routing): replace Arc<RWlock> by RefCell
We store the instructions associated to a next-hop along the
next-hop itself. However, we want to keep next-hops in a BTreeSet,
which requires us to separate the mutable from the immutable
contents of the next-hop. That's achieved by implementing Ord and
Eq and friends. However, BTreeSet does not allow us to iterate
mutably over the next-hops anyway. So, we need interior mutability.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>