test(ci): factor sterile.yml into dev.yml
This uses a matrix expression to obviate the entire sterile.yml workflow.

This significantly reduces code duplication and increases "fail fast" in several ways.

1. The sterile-release build builds the container as soon as possible.  This should reduce break/fix cycles when developing on actual hardware.
2. The non-sterile-debug build is quite fast and should show test failures quickly.  It is, of course, still possible that tests will fail in other branches, but the odds are good that if the non-sterile debug build is successful then the other builds will also be successful.
3. The fuzz build skips pushing a contianer entirely.  This container is not useful for production or testing, so no need to build it.
4. This design has significantly more parallelism (at the expense of downloading more copies of the compile-env).

The longest build (on average) is the sterile-release build, which takes between 5 and 7 minutes.  This is fairly close to the prior sterile.yml run time.  At the same time, we should get actionable results more quickly for most cases.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>