test: add support for "full" fuzz testing
Why I did this
--------------

Running the test suite via `just cargo test` or `just cargo nextest run` will run the fuzz tests, but it does so under significant limitations.

1. The tests (even the fuzz tests) are only run briefly.
2. Coverage information and sanitizers are not enabled.
3. A full fuzzing engine is not set up, so evolutionary feedback is not provided when the tests are run this way,

Using libfuzzer can remove these limitations in exchange for adding additional limits.

1. The tests are run one at a time and can run for much longer (actually, they run as long as you want them to).
2. Coverage information and sanitizers (as well as overflow checks and debug asserts) are enabled.
3. Evolutionary feedback!

The major downside is that these are very computationally heavy processes and can take a long time to run.

How to use what I did
---------------------

To run a full fuzz test, start by listing the available fuzz targets:

  `just list-fuzz-tests`

Then pick a target, e.g. `vxlan::test::mutation_of_header_preserves_contract`, and run `libfuzzer` like so

  `just _test_type=FUZZ fuzz vxlan::test::mutation_of_header_preserves_contract`

The test will run for 1 minute by default, but you can change to, e.g., 15 minutes via

  `just _test_type=FUZZ fuzz vxlan::test::mutation_of_header_preserves_contract -T 15min`

What I did
----------

This is a somewhat tricky and subtle commit, so buckle up.

There were a few closely related issues preventing full fuzzing from working before this change.

The main issue is that `cargo-bolero` (partially) eats our `RUSTFLAGS`.
As a result, we don't have control over the `RPATH` or version of the C++ standard library which is linked into our test executables.
This may seem crazy, but it isn't: the point of bolero is to make fuzzing easy.
It takes control of some linking flags as a result.
Hard to argue with that logic :shrug:.

That said, it is a problem in this case for three reasons.

1. Until recently, we were building all of our `dpdk-sys` dependencies with LLVM's libcxx rather than `libstdc++`. These .so files are absolutely not compatible, and I can't tell bolero to use libcxx from our sysroot (as far as I can tell).
   This fix to this is found in a recent commit to dpdk-sys (a776e846b0edf0f2714f984dc67a25485ccb8714) where I swapped us to `libstdc++` across the board.
   It doesn't really matter which version of the C++ standard library we end up linking against; we don't ship with that library linked at all; it is only necessary for libfuzzer (which is written in C++ because it is part of LLVM).

2. We need to ensure that we are linked against the correct version of `libstdc++`; we want the `libstdc++` from the compile-env, not something from the host system. Again, there is no way to tell bolero which one to use.
   This answer is solved via brute force: use the `compile-env` recipe to run the build for the fuzz profile.
   That way the only version of `libstdc++` available is also the correct version.

3. Bolero does not respect the compile-env's `/nix` `RPATH` (because it doesn't know about it).
   This means that even after solving the first and second problems, we still aren't quite ready to run; we still need to run `patchelf` on the test executable to point the dynamic linker to the canonicalized path to `libstdc++.so` in the compile-env.
   My gut says that simply setting `LD_LIBRARY_PATH` should work, but I was not able to get that to function.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>