test: add support for code coverage tests
To use:

If you run

```
just cargo install cargo-llvm-cov
```

and then run

```
just coverage
```

You will generate coverage reports (html and json) in the target/nextest/coverage directory.

Three reports are generated for each run.

The first is a simple html report.  Start a webserver in that directory to browse results.

The second report is the llvm json coverage format.  Some IDEs can consume this format and display coverage data in the gutters.

The third report is in the format required for uploading to the codecov service.

Coverage tests are run under the fuzz profile to enable coverage related RUSTFLAGS.

A future commit will add support to the CI workflow for uploading coverage data to codecov.

Note:

To make this work I needed to adjust the test-runner.sh script to allow
write access to the target directory for containerized tests.
If you don't do that then the coverage data can't be written out and you
get spurious results indicating that large swaths of the project have
no test coverage.

Additionally, we need to forward a few environment variables set by
cargo-llvm-cov to the containerized environment to tell it where it
should write the coverage data.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>