feat(nat): Make allocator and session manager use a NatIp trait
Instead of using an IpAddr object, make the IP/port allocator and the
session manager use a sealed NatIp trait, to avoid multiple matches for
the exact IP version in the core of the NAT logics.

We need to pass two distinct types to the stateful_nat() function: one
is the type for the original IP address, the second one for the target
IP address.

Suggested-by: Daniel Noland <daniel@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>