feat(routing): reorganize CPI and add lock
Reorganize CPI code defining a CPI object that includes a poller,
the cpi and cli socks and a control channel. Also, add a lock and
unlock control messages to allow locking/unlocking to disallow/
allow updates on the routing database. Explicit unlocking is not
needed since a lockguard is defined to automatically unlock the
CPI when the guard goes out of scope in a RAII fashion. This is
far less error-prone than explicit unlocking.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>