chore(codecov): Add Codecov config file to disable comments in PRs
Comments in Pull Requests discussions from the Codecov bots are
intrusive, in the sense that they break the flow of the conversation,
and they also generate email notifications to all contributors
subscribed to the Pull Request.

Let's disable the comments. The status check at the bottom of the Pull
Requests are enough to get some signal about the change in coverage.

Link: https://docs.codecov.com/docs/pull-request-comments#disable-comment
Signed-off-by: Quentin Monnet <qmo@qmon.net>