feat(routing): handle vrf deletions
When a VRF is to be deleted (due to a config change), we mark it
as Deleting. From that point on, when routes get deleted from it
we check if it becomes empty. When that happens, we change its
status to 'Deleted' so that it can be removed. That is a vrf can
be in 3 states:
  - active: default state when created
  - deleting: we may have removed its vni and fib access and
    keep it to handle deletions.
  - deleted: a vrf being in state deleting that has no more routes
    and can be deleted.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>