feat(routing): delete vrf upfront if empty
Upon receiving a new configuration, if a vrf has to be deleted,
we mark it a Deleting to give a chance to frr to delete the routes.
However, it may be that the vrf is already empty. Check if that is
the case and, if so, remove the vrf without further delay.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>