fix(mgmt): change validation criteria for vtep
We were requiring that a vtep interface always be present. While
this is nearly always the case, it prevents us from pushing
underlay-only configs or empty configs (e.g. to wipe out). Relax
the requirement by allowing "one vtep at the most, and only if
needed." We determine the latter by the presence of vpc configs.

Reuse method to address the above when building the config.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>