refactor(nat): Make NatSession a trait
To make it easier to test alternative session managers, turn struct
NatSession into a trait so that implementations of the NatSessionManager
trait are not dependent on a particular NatSession structure. Trait
NatSessionManager becomes generic, able to work with any underlying
struct that implements trait NatSession.

Provide a default implementation NatDefaultSession, to be used with the
NatDefaultSessionManager. Implementation will be completed in future
commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>