feat(nat): Add sessions to Nat; Implement sessions lookup, creation
Add a NatDefaultSessionManager object to the Nat pipeline stage, and
implement lookup and (partial) session creation for this session
manager - partial because the reverse session creation is missing at
this point.

In the future, the Nat object could trivially be made generic related to
the NatSessionManager implementation in use. For now, we only have the
default one, so use it unconditionally.

Signed-off-by: Quentin Monnet <qmo@qmon.net>