feat(nat): Implement a dashmap-based NatSession and NatSessionManager
Use dashmap to implement our NatDefaultSessionManager. We use two maps
to store entries, one with IPv4 and one with IPv6 for NatTuple lookups,
and the values are IpAddr so that we can look retrieve either an IPv4 or
an IPv6 for any given address, depending on whether we're doing NAT44 or
another variant in the future.

Implement the different functions for this session manager. Note that
because of the concurrency aspects for the dashmap entries, we need to
add some lifetime parameters to several objects. We also make our
NatDefaultSession generic related to the IP version of the associated
NatTuple, because we need to make the reference to the dashmap generic,
and this reference contains the whole map entry - the NatState and the
associated NatTuple<I: NatIp>.

Signed-off-by: Quentin Monnet <qmo@qmon.net>