feat(nat): Add a (yet non-functional) allocator to struct Nat
Rework the skeleton that we had for the NAT allocator. Make a
NatAllocator trait, implement a default allocator (just like we did for
the NAT session managers). Add one to struct Nat.

Make allocate() a method of the allocator, rather than having the
"NatPool" indirection to first retrieve a pool and then an IP/port.
Finding the right pool, if necessary, can remain internal to the
allocator.

Signed-off-by: Quentin Monnet <qmo@qmon.net>