fix(mgmt): parse and store AF networks to advertise
gateway_config 0.10.1 had the wrong model for static advertisements
gateway_config 0.11.0 fixes the model

This commit adds networks to address families and removes them from
BgpNeighbor to match the new model.  It also fixes the gRPC converter
to properly store and serialize the address family networks.

This commit does *not* change the frr renderer to output the
config needed to actually advertise the networks.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>