feat(net): do not compute udp checksum for vxlan
Do not recompute udp checksum when packet has vxlan encapsulation.
If we create the encapsulation, the checksum is initialized to 0,
which is fine since receiving endpoints must be able to decapsulate
it and needs not be recalculated. If we're not creating the
encapsulation, that means that we did not decapsulate the packet,
in which case, there should not be the need to recompute it since
the pseudo-header did not change.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>

wip