ci: Use "release" environment for fuzzing
In a recent commit we adjusted the build and CI environments to use
either the release or the debug images for the compile-env and libc-env
containers. We made the "release" profile use the "release" environment,
the "debug" profile use the "debug" environment, as one might expect. As
for the "fuzz" profile? We made it use the "debug" environment, but this
was the wrong pick: it should be "release", because we're using the same
LTO parameters for fuzzing as with the "release" environment. This
commit makes "release" the default environment used with everything
other than "debug", so that the "fuzz" profile gets the correct
environment.

Fixes: fa9137378108 ("fix(ci): Update compile-env container name after dpdk-sys changes")
Signed-off-by: Quentin Monnet <qmo@qmon.net>