build: Fix warning for compile-env presence check
Following recent changes in dpdk-sys and in the dataplane repository,
the container names now have a ".debug" or ".release" suffix. We forgot
to update the compile-env check accordingly. There's no reason to
re-create the full container path in the warning, by the way: let's just
use _compile_env_container that contains the full string already.

Fixes: fa9137378108 ("fix(ci): Update compile-env container name after dpdk-sys changes")
Reported-by: Sergey Matov <sergey.matov@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>