feat(mgmt): disable interface config in tests
The test test_sample_config() turns out to exercise a significant
chunk of the management/config logic, even if it uses a fake
frr-agent. Specifically, it exercices the building of the internal
config, the rendering of the frr config, the vpc-manager and
creation of kernel interfaces and the creation and application of
the router config. Currently, however, the interface manager
does not create kernel interfaces for the physical ports. As a
result, the configuration of interfaces in the router fails as
these are not found in the test container.

To fix this, temporarily inhibit the interface configuration when
running the test.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>