feat: introduce tc filter
A filter is basically a tuple of

1. a match criteria; some list of selectors which decide if this filter applies to a given packet.
2. a list of actions; some set of operations to perform if the match criteria apply.
3. a priority; a number to disambiguate which filter applies if multiple filters match.  In linux, tc filters match _lower_ priority first.

Filters exist within a chain, and chains live on network interfaces or blocks of network interfaces.

Some examples of filters include matching on ARP, or IPv4 packets with source ip in the range `192.168.5.0/24`.
Match criteria can be combined to form complex filters.
In such cases, a packet must match all listed criteria to trigger the actions.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>