feat(mgmt): filter-out adv of host routes
Do not advertise host routes if possible. This is an optimization
that can help significantly save resources in the fabric nodes and
reduce bgp churn. The method used is not infallible because it
trades off effectiveness with complexity in the gateway
configuration. Specifically, if a VPC has peerings that have host
prefixes, it may not filter other host routes within other non-host
prefixes. More elaborate methodologies may allow for that at the
cost of higher complexity and significantly higher processing of
BGP routes.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>