feat(nat): Implement PartialEq, Eq for NatTables (and related types)
Implement PartialEq and Eq traits for struct NatTables, so we can
compare different objects in tests in a future commit.

We need to dreive the traits for structs PerVniTable,
NatPrefixRuleTable, NatPeerRuleTable, and TrieValue. We also need the
trait to be implemented for PrefixTrie, but we cannot derive it (the
RTrieMap objects do not implement the trait), so we implement it
manually by comparing all entries.

Signed-off-by: Quentin Monnet <qmo@qmon.net>