test(nat): Add stateless NAT test in "stateless" module
We used to have a similar test in the module, but we recently moved it
to the "mgmt" crate because it depended on some objects from "mgmt", but
the "nat" crate can no longer use "mgmt" as a dependency.

Restore the test in the stateless NAT submodule, but instead of building
the NAT tables from the configuration code from the "mgmt" crate, just
build the table manually.

In a future commit, we'll remove the copy of the test from the "mgmt"
crate and will instead check that the generated config object matches
with the manually-initialised NatTables object from this commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>