test(mgmt): Remove redundant tests for stateless NAT
In a recent commit we copied the tests for stateless NAT from the "mgmt"
crate back to the "nat" crate. We no longer need the version from the
"mgmt" crate. Instead, ensure that the NatTables object generated from
the config matches the reference object that we manually instantiated in
the tests of the "nat" crate. In order to share and compare this object,
gate it behind a "testing" feature for the "nat" crate.

Signed-off-by: Quentin Monnet <qmo@qmon.net>