test(net): Use distinct src and dst IPs in build_test_ipv4_packet()
It's better to have distinct source and destination IP addresses in the
test packet that we generate with build_test_ipv4_packet(), to avoid any
confusion as to what address we're processing (and because it makes
little sense to use the same addresses in the first place).

Adjust related tests accordingly: these are the tests for stateless NAT,
in the "mgmt" package.

Suggested-by: Fredi Raspall <fredi@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>