fix(nat): Fix source NAT tables build
We've been using the wrong prefixes when building the peers table for
the first lookup for source NAT: we've used the "ips" blocks instead of
the "as_range". We missed it in the tests, because we also use the wrong
address for the lookup: we would do source IP when we should use the
destination IP to find the remote VPC, of course.

Fix both the table and the lookup.

Reported-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>