feat(net,dataplane): annotate if need nat
Add a boolean that indicates if the nat stage should process a
packet or not. This eases the logic of the nat module and prevents
attempting to nat packets that shouldn't be nated. Note that the
fact that a packet is marked as to be natted does not mean that
any of its addresses may change, since the nat stage may conclude
there's no need to.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>